Fix compiler warning

+++ ares/ares__read_line.c
@@ -1,4 +1,4 @@
-/* $Id: ares__read_line.c,v 1.11 2009-10-08 18:38:29 yangtse Exp $ */
+/* $Id: ares__read_line.c,v 1.12 2009-10-09 14:06:38 yangtse Exp $ */
 
 /* Copyright 1998 by the Massachusetts Institute of Technology.
  *
@@ -46,7 +46,7 @@ int ares__read_line(FILE *fp, char **buf, size_t *bufsize)
 
   while (1)
     {
-      if (!fgets(*buf + offset, *bufsize - offset, fp))
+      if (!fgets(*buf + offset, (int)(*bufsize - offset), fp))
         return (offset != 0) ? 0 : (ferror(fp)) ? ARES_EFILE : ARES_EOF;
       len = offset + strlen(*buf + offset);
       if ((*buf)[len - 1] == '\n')
